Implement Metal backend for inference service
- Added Metal compute kernel for probing functionality in `probe.metal`.
- Developed `MetalDevice` class to manage Metal device and command queue, including capabilities querying and shader compilation.
- Created `MetalInferenceService` to handle inference requests, manage sessions, and process requests concurrently.
- Implemented C API smoke test to validate Metal backend integration.
- Added unit tests for Metal device and inference service to ensure functionality and performance parity with CPU implementation.
- Introduced quantization tests to verify Metal's handling of quantized tensors against CPU reference.

+++ src/backend/cpu/weights.cpp
@@ -26,7 +26,8 @@ std::string source_identity(const std::filesystem::path& path) {
     if (std::filesystem::is_regular_file(path)) {
         out << std::filesystem::file_size(path) << ':';
     }
-    out << std::filesystem::last_write_time(path).time_since_epoch().count();
+    out << static_cast<long long>(
+        std::filesystem::last_write_time(path).time_since_epoch().count());
     return out.str();
 }
